Refactor crop button
